+++ internal/models/budget.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Budget struct {
 	gorm.Model
 	Name       string
-	Amount     float64
+	Amount     float32
 	CategoryId uint
 	Category   Category
 }
